add human readable standard_used and standards_supported key values even though xtse is supported
Add the human readable texts to standard used and supported for Intel platform. The major difference between
XTSE and old standards is that the former has smaller granularity than the latter. Now all sub-standards will
be converted to the one single standard.

The problem of inconsistent status of line and link is also fixed. Force the line's status to be down if the
link is not up.